Increase timeout for detecting a reset connection in the integration test
The part of the integration test which tests for reset connections after
a key is revoked is flaky on Travis.  We sometimes get an error code 52
isntead: "The server didn't reply anything, which here is considered an
error."

As a first attempt to deflake this test, we increase the connection
timeout for the request.  5 seconds was chosen completely arbitrarily.