Consistent orientation for the Outline logo (#571)
Maintaining consistency with https://github.com/Jigsaw-Code/outline-client/pull/743 as well as with the app logos for both the client and the manager, flip all of the Outline discs so the light side is on the right.  I didn't touch a few (such as metrics.png) which are more complicated and would require access to the file in whatever image editor was used to create them.