Remove LoadingWindow (#734)
From conversations with @alalamav, at one point the main window needed
connect to the Internet and so we would display a loading page if it
took more than 3 seconds to load.  The main window no longer accesses
the internet, so there's no real reason for us to keep a loading page
around which won't be displayed.