Preserve server creation in-progress state
Currently, if the user exits the GCP "app" after triggering a
server creation in the region picker, and returns, they will see
a fresh region picker, with no indication that there is an ongoing
server creation action (until server creation succeeds).

This change preserves the server creation UI state, so that users
see the ongoing server creation attempt when they return to the
GCP "app".