GCP: Make server creation more responsive
Currently, if the user starts server creation in GCP, they are stuck
on the region picker page until server creation completes. This takes
a long time, and is different from DigitalOcean, where server creation
is faster because it hasn't fully completed when the creation request
returns.

This change makes the GCP behavior match the DigitalOcean behavior, by
creating a ManagedServer object as soon as we know what the instance's
ID will be. This works by creating the Server object earlier, before
the instance is actually ready.