Migrate all UI components to Typescript
Our UI components are in a mixture of

* Polymer 1.0 style in JS
* Polymer 2.0 style in JS
* Polymer 3.0 style in TS
* Lit-Element 1.0 in JS
* Lit-Element 2.0 in TS

This change migrates all JS to TS, leaving us with slightly less
divergence between component implementations.  This change only adds
type annotations and corrects bugs encountered during type checking.
Some notable bugs fixed:

* We were annotating properties "readonly" throughout, but the correct
annotation is "readOnly".  However, the effect of this annotation is not
what we thought it was (it actually makes any attempt to set or bind the
property fail silently, even in the constructor), so I've removed most
uses of this annotation.
* Missing imports and dependencies.
* Several instances of type confusion.

Hopefully, migrating to Typescript will make the code easier to
navigate with automated tools, especially for new team members, and will
accelerate subsequent cleanups.