Fix login bug after free trial period expires
When the free trial period expires, the Outline project is left linked
to an inactive billing account, and all instances are deleted.
Attempting to list the instances associated with a project that lacks
billing results in an HTTP 403.  Currently, this results in GCP login
failure, with a generic error message.

This change makes the manager handle this account state gracefully,
taking the user to the billing setup screen, and from there directly
to the server creation screen.