fix(manager): Proxy Shadowbox HTTP requests through Node
Currently, the Outline Server manager accesses Shadowbox APIs
using the browser's `fetch()` capability.  To allow the use
of self-signed certificates, the main process maintains a
list of trusted self-signed certificates and overrides the
browser's certificate verification errors when these
certificates are received.  This works well for a single
server, but creates potential security concerns when multiple
servers are enabled, because these servers necessarily share
the global trusted certificate list, as well as other state in
the browser network stack (e.g. the socket pool).

This change moves all communication with the Shadowbox server
to the Node.js HTTP client, which enables detailed control over
TLS configuration and does not rely on any shared state.

This is an alternative to #1086.