fix(manager): 💊 integrate with the latest Sentry library (#1309)
In this PR, I integrated Outline Manager with the latest Sentry library and re-enabled the user feedback form. Sentry was not working because of two reasons:
1. Electron enabled sandbox to its `preload.js` script, therefore calling `Sentry.init` in `preload` is not enough (and not necessary), we must call `Sentry.init` in every renderer process as well
2. `SENTRY_DSN` is not injected into electron, we read it from `process.env.SENTRY_DSN`, but this is a runtime environment variable (which will be `undefined`)

So basically my fix is straightforward:
1. Calling `Sentry.init` in `web_app/app.ts`, which is the entrypoint of renderer process
2. Use `webpack` to inject `SENTRY_DSN_` variable into `electron_app/index.ts` during build

In order to be consistent (and to minify the result js file), I also introduced webpack to build the `preload.ts` script as well.

Finally I upgraded electron to the latest version that does not have [API restrictions](https://www.electronjs.org/docs/latest/breaking-changes#default-changed-renderers-without-nodeintegration-true-are-sandboxed-by-default) to `preload` script (we are using `crypto` package which will be forbidden). A refactoring is required in order to upgrade to any newer versions of electron.