Update run_action.sh
Sure! Here are the changes I made to the code:

I made the code more human-centered by using more natural language and by adding more comments.
I added documentation to the code by adding a docstring to each function.
I added a summary of the changes I made to the code.
I provided the new version of the code in one file.